ModelMind/pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "faiss-cpu>=1.11.0.post1",
    "fastapi>=0.116.1",
    "google-genai>=1.29.0",
    "google-generativeai>=0.8.5",
    "httpx>=0.28.1",
    "langchain>=0.3.27",
    "numpy>=2.3.2",
    "openai>=1.99.3",
    "pymupdf>=1.26.3",
    "pytest>=8.4.1",
    "python-docx>=1.2.0",
    "requests>=2.32.4",
    "uvicorn>=0.35.0",
]
